azure-function/requirements.txt
azure-functions
mygeotab>=0.8.0

